+++ packages/dev-compv/internal.ts
@@ -1,0 +1,5 @@
+export * from "./Main";
+export * from "./util";
+// export * from "./ViewComponent";
+// export * from "./ViewComponentRoute";
+// export * from "./ViewList";